requirements.txt
Flask
flask_cors
Gunicorn
numpy
matplotlib
pillow
pyocr
tesseract